requirements.txt
kafka-python
fastapi
uvicorn
elasticsearch